updatingEmployeeRole functions completed and updates the employer role
