+++ public/options/adding-data-functions.js
@@ -125,43 +125,36 @@ const positions = await getListOfRoles();
 
 }
 
-const updatingEmployeeRole = async(table = 'employee')=>{
+// const updatingEmployeeRole = async(table = 'employee')=>{
 
-    const positions = await getListOfRoles();
-    const workers = await getListOfEmployees()
+//     const positions = await getListOfRoles();
+//     const workers = await getListOfEmployees()
 
-    const response = await inquirer.prompt([
+//     const response = await inquirer.prompt([
 
-        {
-            type:"list",
-            message:"Which employee would you like to update?",
-            name:"employee",
-            choices:workers
-        },
-        {
-            type:"list",
-            message:"SELECT",
-            name:"role",
-            choices:positions
+//         {
+//             type:"list",
+//             message:"Which employee would you like to update?",
+//             name:"employee",
+//             choices:workers
+//         },
+//         {
+//             type:"list",
+//             message:"SELECT",
+//             name:"role",
+//             choices:positions
     
-        },]);
+//         },]);
 
-        console.log(table);
+//         console.log(table);
 
-        console.log(response);
-
-
+//         console.log(response);
 
 
-
-}
+// }
 
 
 
 
 
-
-
-
-
-module.exports = {addingDepartment,addingRole, addingEmployee,updatingEmployeeRole}
+module.exports = {addingDepartment,addingRole, addingEmployee}
